move Vector3f into separate file
